adding a `bcn_before_loop` filter hook
This allows to do modifications in a similar way that `bcn_after_fill` does, but apply them only to a particular render of the breadcrumbs. That is especially useful if you need to do something different when you have more than one instance of breadcrumbs of the page, i.e. one at the top and one at the bottom.